requirements.txt
pydantic-ai>=0.0.1
httpx>=0.26.0
rich>=13.7.0
logfire>=0.8.0
firecrawl-py>=0.1.0
python-dotenv>=1.0.0
streamlit>=1.29.0 